Begin adding flask api
