+++ docker-compose.yml
@@ -2,17 +2,20 @@ version: "3"
 
 services:
   web:
+    container_name: web
+    build:
+      context: ./services/web
+      dockerfile: Dockerfile
     image: jenngeorge/bikeshare_app:flask_app
-    # deploy:
-    #   replicas: 3
-    #   resources:
-    #     limits:
-    #       cpus: "0.1"
-    #       memory: 50M
-    #   restart_policy:
-    #     condition: on-failure
+    volumes:
+      - ./services/web:/usr/src/bikeshare_app
+    restart: always
     ports:
       - "4000:80"
+    environment:
+      - FLASK_APP=api/__init__.py
+      - FLASK_DEBUG=1
+      - APP_SETTINGS=api.config.DevelopmentConfig
 
   spark-master:
     container_name: spark-master
