+++ services/web/api/config.py
@@ -1,0 +1,19 @@
+
+class BaseConfig:
+    """Base configuration"""
+    TESTING = False
+
+
+class DevelopmentConfig(BaseConfig):
+    """Development configuration"""
+    pass
+
+
+class TestingConfig(BaseConfig):
+    """Testing configuration"""
+    TESTING = True
+
+
+class ProductionConfig(BaseConfig):
+    """Production configuration"""
+    pass
