+++ services/web/manage.py
@@ -1,0 +1,9 @@
+from flask.cli import FlaskGroup
+from api import app
+
+
+cli = FlaskGroup(app)
+
+
+if __name__ == '__main__':
+    cli()
