+++ services/web/requirements.txt
@@ -1,1 +1,3 @@
 Flask==1.0.2
+Flask-SQLAlchemy
+psycopg2
